IEnvironmentService#userHome is deprecated Changes for simplefiledialog.ts Part of #94506
