+++ src/vs/code/electron-browser/sharedProcess/contrib/languagePackCachedDataCleaner.ts
@@ -10,7 +10,7 @@ import product from 'vs/platform/product/common/product';
 import { Disposable, toDisposable } from 'vs/base/common/lifecycle';
 import { onUnexpectedError } from 'vs/base/common/errors';
 import { ILogService } from 'vs/platform/log/common/log';
-import { INativeEnvironmentService } from 'vs/platform/environment/node/environmentService';
+import { INativeEnvironmentService } from 'vs/platform/environment/common/environmentService';
 import { IEnvironmentService } from 'vs/platform/environment/common/environment';
 
 interface ExtensionEntry {
