+++ src/vs/platform/extensionManagement/node/extensionTipsService.ts
@@ -8,7 +8,7 @@ import { join, } from 'vs/base/common/path';
 import { IProductService, IExeBasedExtensionTip } from 'vs/platform/product/common/productService';
 import { IEnvironmentService } from 'vs/platform/environment/common/environment';
 import { env as processEnv } from 'vs/base/common/process';
-import { INativeEnvironmentService } from 'vs/platform/environment/node/environmentService';
+import { INativeEnvironmentService } from 'vs/platform/environment/common/environmentService';
 import { IFileService } from 'vs/platform/files/common/files';
 import { isWindows } from 'vs/base/common/platform';
 import { isNonEmptyArray } from 'vs/base/common/arrays';
